ai/requirements.txt
# Python Compatibility: 3.10 or 3.11 recommended for all dependencies below

# Core FastAPI and Web Framework
fastapi
uvicorn
pydantic>=2.10.0,<3.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.32.3
python-multipart>=0.0.9

# gRPC and Protocol Buffers
grpcio>=1.66.0
grpcio-tools>=1.66.0

# AI/ML Core Libraries
langchain-groq

# LangChain Ecosystem
langchain>=0.3.0

# Embeddings and Vector Databases
sentence-transformers>=3.0.0
chromadb>=0.5.0
faiss-cpu>=1.9.0
qdrant-client>=1.12.0

# Data Processing and ML
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
transformers>=4.46.0
torch>=2.5.0
tiktoken>=0.7.0

# Caching and Task Queue
redis>=5.2.0
celery>=5.4.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Search and Database
elasticsearch>=8.15.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
cryptography>=44.0.0
pyopenssl>=24.3.0

# Observability and Monitoring
prometheus-fastapi-instrumentator>=7.0.0
python-json-logger>=2.0.7

# OpenTelemetry (Stable for Python 3.10/3.11)
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-http>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-exporter-prometheus>=0.38b0,<1.0.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-asgi>=0.48b0
opentelemetry-instrumentation-logging>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
opentelemetry-instrumentation-redis>=0.48b0
opentelemetry-instrumentation-elasticsearch>=0.48b0
opentelemetry-instrumentation-sqlalchemy>=0.48b0

# ELK Stack Integration
elastic-apm==6.24.0
loguru>=0.7.2

# Rate Limiting
slowapi>=0.1.9
limits>=3.13.0

# System Monitoring
psutil>=6.0.0
